fix(*): small cleanups
