+++ app/lib/types/document.ts
@@ -7,3 +7,15 @@ export interface Document {
   inserted_at?: string;
   updated_at?: string;
 }
+
+export interface Collection {
+  id: string;
+  name?: string;
+  description?: string;
+  knowledge_base_id?: string;
+  inserted_at: string;
+  updated_at: string;
+  status?: string;
+  job_id?: string;
+  documents?: Document[];
+}
